requirements.txt
requests==2.5.3
Fabric==1.9.1
PyMySQL==0.9.2
Django==1.11.16
django-celery-beat==1.4.0
django-celery-results
django-cors-headers==2.4.0
django_slack==5.8.0
djangorestframework==3.9.0
httplib2==0.10.3
mechanize==0.2.5
beautifulsoup4==4.5.1
pycrypto==2.6.1
gunicorn==19.9.0
celery==4.2.2
flower==0.9.3
django-environ==0.4.5
uWSGI==2.0.18
redis==3.3.11
sklearn==0.0
pandas==0.24.2
django-picklefield==2.0